all: re-license CrosEC et al to MIT
I don't have line of sight on whether a GPLed kernel-mode driver, with
required code signing, can be distributed through Windows Update.

If CrosEC were signed and distributed as part of a system image or
shipped through Windows Update, a user couldn't necessarily replace it
with a like-working copy due to code signing without putting their
device in test mode ... thus restricting a freedom granted by the
license.

Pursuant to section 6 of the GPL, which includes "...You may not impose
any further restrictions on the recipients' exercise of the rights
granted herein." I do not believe that this project *can* function under
the GPL.

Anyway, this project is only a derived work of the Chromium EC which is
also released under the terms of the 3-clause BSD license. It bears a
passing similarity to the Linux cros-ec-lpcs driver in *function*, but
not in *form*. This is required for interoperability.

Signed-off-by: Dustin L. Howett <dustin@howett.net>