fix: enforce minimum required command length in CrosECIoctlXCmd (#10)
Instead of the minimum required command length being the size of the
struct `sizeof(CROSEC_COMMAND)`, the size of the **pointer** to the
struct `sizeof(PCROSEC_COMMAND)` was being used.

For this to have a real world impact, the userland code would have to
send a command to the driver whose input buffer is too small. Then
`CrosECIoctlXCmd` would read past the input buffer.